Add sdl-generic-hmi.bb (generic_hmi)
NOTE: The required files for the HMI are only index.html and
build/bundle.js.  Those are provided by the git repository and simply
installed by this recipe, but they seem to be originally created out of
some build process using node.js & webpack(?)  I want to note that this
bitbake recipe is NOT including any rebuild from source material since
the result files have been provided, and since I'm not sure how that is
intended to work for this repository. This approach serves our purposes
for now.

Signed-off-by: Gunnar Andersson <gandersson@genivi.org>
Signed-off-by: Oscar Andreasson <oscar.andreasson@pelagicore.com>